% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors.R
\name{get_collectors}
\alias{get_collectors}
\title{get_collectors}
\usage{
get_collectors(
  survey_id,
  page = 1,
  per_page = NULL,
  sort_by = NULL,
  sort_order = NULL,
  name = NULL,
  start_date = NULL,
  end_date = NULL,
  include = NULL,
  all_pages = TRUE,
  oauth_token = get_token()
)
}
\arguments{
\item{survey_id}{the id of the survey whose collectors you want,
find it with \code{\link{browse_surveys}}.}

\item{page}{Integer number to select which page of resources to return. By default is 1.}

\item{sort_by}{column to sort by. Can be one of: id, date_modified, type, status, name.}

\item{sort_order}{sort data by the value in `sort_by`. Can be one of: ASC, DESC.}

\item{name}{Nickname of collector to search against}

\item{start_date}{Collectors must be created after this date. Date string in format YYYY-MM-DDTHH:MM:SS (no offset)}

\item{end_date}{Collectors must be created before this date. Date string in format YYYY-MM-DDTHH:MM:SS (no offset)}

\item{include}{Specify additional fields to return per collector. Can be one or more of: type, status, response_count, date_created, date_modified, url. Optionally, specify "everything" to return all possible fields.}

\item{all_pages}{return all pages of respondents?
Default is TRUE, which will fetch all responses (and cause n/100 calls to the API).}

\item{oauth_token}{Your OAuth 2.0 token.
By default, retrieved from \code{get_token()}.}
}
\value{
a data.frame (technically a \code{tibble}) with each collector and its information.
}
\description{
Get collectors for a SurveyMonkey survey
}
