% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_questions.R
\name{get_questions}
\alias{get_questions}
\title{survey_questions}
\usage{
get_questions(pc)
}
\arguments{
\item{pc}{a survey details object, the result of a call to fetch_survey_details}
}
\value{
A data frame with one row per question/subquestion/answer choice
}
\description{
Creates a data frame from the survey questions and answers
}
