% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_responses.R
\name{get_responses}
\alias{get_responses}
\title{get_responses}
\usage{
get_responses(
  id,
  page = 1,
  all_pages = TRUE,
  per_page = 100,
  start_created_at = NULL,
  end_created_at = NULL,
  start_modified_at = NULL,
  end_modified_at = NULL,
  sort_order = "DESC",
  sort_by = "date_modified",
  oauth_token = get_token(),
  verbose = TRUE
)
}
\arguments{
\item{id}{The survey's ID, can be found with \code{browse_survey()}.}

\item{page}{Integer number to select which page of resources to return. By default is 1.}

\item{all_pages}{return all pages of respondents?
Default is TRUE, which will fetch all responses (and cause n/100 calls to the API).}

\item{per_page}{Integer number to set the number of surveys to return per page.
By default, is 100 surveys per page (appears to be the maximum allowed by the API).}

\item{start_created_at}{Date string used to select surveys created after this date.
By default is NULL.}

\item{end_created_at}{Date string used to select surveys modified before this date.
By default is NULL.}

\item{start_modified_at}{Date string used to select surveys last modified after this date.
By default is NULL.}

\item{end_modified_at}{Date string used to select surveys modified before this date.
By default is NULL.}

\item{sort_order}{String used to set the sort order for returned surveys:
'ASC’ or 'DESC’. By default, DESC.}

\item{sort_by}{String value of field used to sort returned survey list:
‘title’, 'date_modified’, or 'num_responses’. By default, date_modified.}

\item{oauth_token}{Your OAuth 2.0 token.
By default, retrieved from \code{get_token()}.}

\item{verbose}{Show API rate limit messages?}
}
\value{
A list of object of class {sm_response}
}
\description{
Get responses for a SurveyMonkey survey
}
\references{
SurveyMonkey API V3 at
\url{https://developer.surveymonkey.com/api/v3/#survey-responses}
}
