% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{parse_responses}
\alias{parse_responses}
\title{Take a survey object and parses it into a tidy data.frame.}
\usage{
parse_responses(surv_obj)
}
\arguments{
\item{surv_obj}{a survey, the result of a call to \code{fetch_survey_obj}.}
}
\value{
a data.frame (technically a \code{tibble}) with clean responses, one line per respondent.
}
\description{
Take a survey object and parses it into a tidy data.frame.
}
