% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble.R
\name{parse_survey}
\alias{parse_survey}
\title{Take a survey object and parses it into a tidy data.frame.}
\usage{
parse_survey(
  surv_obj,
  oauth_token = get_token(),
  ...,
  fix_duplicates = c("error", "drop", "keep", "none")
)
}
\arguments{
\item{surv_obj}{a survey, the result of a call to \code{fetch_survey_obj}.}

\item{oauth_token}{Your OAuth 2.0 token. By default, retrieved from
\code{get_token()}.}

\item{...}{additional arguments to pass on to \code{get_responses}.  See the documentation
\code{?get_responses} where these arguments are listed.}

\item{fix_duplicates}{character if 'error', the default detection of duplicate data will result
in an error being raised, otherwise allow the function to return. If 'keep' duplicate results
will be retained, if 'drop' duplicates will be removed from the results.

If encountering issues with duplicates, use `fix_duplicates = 'none'` to deal with duplicate responses after
the survey is parsed.}
}
\value{
a data.frame (technically a \code{tibble}) with clean responses, one line per respondent.
}
\description{
Take a survey object and parses it into a tidy data.frame.
}
