% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_token}
\alias{set_token}
\title{Set Survey Monkey Oauth Token}
\usage{
set_token(oauth_token)
}
\arguments{
\item{oauth_token}{Your survey monkey OAuth 2.0 token.}
}
\description{
Sets the option 'sm_oauth_token' for the survey monkey API.
Token will be used for authorization headers for requests.
}
