% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_utils.R
\name{sm_get}
\alias{sm_get}
\title{Get request for survey monkey API}
\usage{
sm_get(url, query, config, ..., verbose = TRUE)
}
\arguments{
\item{url}{character, url to send request to}

\item{query}{character, components of the url to change for request arguments,
see also \code{?httr::modify_url}, \link[httr]{modify_url}}

\item{config}{see \code{?httr::config}, \link[httr]{config}, for full details.
Additional configuration settings such as http additional headers.}

\item{...}{additional argument passed to \code{httr::GET} \link[httr]{GET}.}

\item{verbose}{logical, if TRUE print some messages about request limits}
}
\description{
Will always set the tntp/surveymonkey repo as the user agent
}
