% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_survey.R
\name{fetch_survey_obj}
\alias{fetch_survey_obj}
\title{Title}
\usage{
fetch_survey_obj(id, oauth_token = getOption("sm_oauth_token"))
}
\arguments{
\item{id}{ID number of survey to be fetched.}

\item{oauth_token}{Your OAuth 2.0 token.
By default, retrieved from \code{getOption('sm_oauth_token')}.}
}
\value{
a survey object, which is a nested list containing info about the survey.
}
\description{
Title
}
\examples{
# not run:
# fetch_survey_obj(123456789)
}
