% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors.R
\name{get_collectors}
\alias{get_collectors}
\title{get_collectors}
\usage{
get_collectors(
  survey_id,
  page = 1,
  all_pages = TRUE,
  oauth_token = get_token()
)
}
\arguments{
\item{survey_id}{the id of the survey whose collectors you want,
find it with \code{\link{browse_surveys}}.}

\item{page}{Integer number to select which page of resources to return. By default is 1.}

\item{all_pages}{return all pages of respondents?
Default is TRUE, which will fetch all responses (and cause n/100 calls to the API).}

\item{oauth_token}{Your OAuth 2.0 token.
By default, retrieved from \code{get_token()}.}
}
\value{
a data.frame (technically a \code{tibble}) with each collector and its information.
}
\description{
Get collectors for a SurveyMonkey survey
}
