% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_questions.R
\name{get_questions}
\alias{get_questions}
\title{survey_questions}
\usage{
get_questions(id)
}
\arguments{
\item{id}{SurveyMonkey survey id.}
}
\value{
A data frame with one row per question/subquestion/answer choice
}
\description{
Creates a data frame from the survey questions and answers
}
